+++ src/calculations.ts
@@ -94,7 +94,7 @@ const easter = (year: number) => {
   const easterMonth = Math.floor((h + l - 7 * m + 114) / 31);
   const easterDay = ((h + l - 7 * m + 114) % 31) + 1;
 
-  return new Date(year, easterMonth - 1, easterDay);
+  return new Date(Date.UTC(year, easterMonth - 1, easterDay));
 };
 
 const findNextWeekDay = (
@@ -103,7 +103,7 @@ const findNextWeekDay = (
   day: number,
   targetWDay: number
 ) => {
-  const date = new Date(year, month, day);
+  const date = new Date(Date.UTC(year, month, day));
   date.setDate(date.getDate() + ((targetWDay - date.getDay() + 7) % 7));
   return date;
 };
@@ -113,8 +113,8 @@ let solsticeBaseSummer: number;
 let solsticeBaseWinter: number;
 const solstice = (year: number, season?: "winter" | "summer") => {
   if (!solsticeBaseSummer) {
-    solsticeBaseSummer = new Date(2016, 5, 20,22,34).getTime();
-    solsticeBaseWinter = new Date(2016, 11,21,10,44).getTime();
+    solsticeBaseSummer = new Date(Date.UTC(2016, 5, 20, 22, 34)).getTime();
+    solsticeBaseWinter = new Date(Date.UTC(2016, 11, 21, 10, 44)).getTime();
   }
   let time = season === "winter" ? solsticeBaseWinter : solsticeBaseSummer;
   time = time + solsticeInterval * (year - 2016);
@@ -131,7 +131,7 @@ export const calcSpecialDays = (year: number) => {
     const easterSundayMs = easterSunday.getTime();
     const _holidays = [
       {
-        date: new Date(year, 0, 1),
+        date: new Date(Date.UTC(year, 0, 1)),
         description: "Nýársdagur",
         key: "nyars",
         holiday: true,
@@ -161,7 +161,7 @@ export const calcSpecialDays = (year: number) => {
         holiday: false,
       },
       {
-        date: new Date(year, 1, 14),
+        date: new Date(Date.UTC(year, 1, 14)),
         description: "Valentínusardagur",
         key: "valent",
         holiday: false,
@@ -203,7 +203,7 @@ export const calcSpecialDays = (year: number) => {
         holiday: true,
       },
       {
-        date: new Date(year, 4, 1),
+        date: new Date(Date.UTC(year, 4, 1)),
         description: "Verkalýðsdagurinn",
         key: "mai1",
         holiday: true,
@@ -227,7 +227,7 @@ export const calcSpecialDays = (year: number) => {
         holiday: true,
       },
       {
-        date: new Date(year, 5, 17),
+        date: new Date(Date.UTC(year, 5, 17)),
         description: "Þjóðhátíðardagurinn",
         key: "jun17",
         holiday: true,
@@ -239,7 +239,7 @@ export const calcSpecialDays = (year: number) => {
         holiday: false,
       },
       {
-        date: new Date(year, 5, 24),
+        date: new Date(Date.UTC(year, 5, 24)),
         description: "Jónsmessa",
         key: "jonsm",
         holiday: false,
@@ -251,13 +251,13 @@ export const calcSpecialDays = (year: number) => {
         holiday: true,
       },
       {
-        date: new Date(year, 9, 31),
+        date: new Date(Date.UTC(year, 9, 31)),
         description: "Hrekkjavaka",
         key: "hrekkja",
         holiday: false,
       },
       {
-        date: new Date(year, 11, 1),
+        date: new Date(Date.UTC(year, 11, 1)),
         description: "Fullveldisdagurinn",
         key: "fullv",
         holiday: false,
@@ -269,32 +269,32 @@ export const calcSpecialDays = (year: number) => {
         holiday: false,
       },
       {
-        date: new Date(year, 11, 23),
+        date: new Date(Date.UTC(year, 11, 23)),
         description: "Þorláksmessa",
         key: "thorl",
         holiday: false,
       },
       {
-        date: new Date(year, 11, 24),
+        date: new Date(Date.UTC(year, 11, 24)),
         description: "Aðfangadagur",
         key: "adfanga",
         holiday: true,
         halfDay: true,
       },
       {
-        date: new Date(year, 11, 25),
+        date: new Date(Date.UTC(year, 11, 25)),
         description: "Jóladagur",
         key: "jola",
         holiday: true,
       },
       {
-        date: new Date(year, 11, 26),
+        date: new Date(Date.UTC(year, 11, 26)),
         description: "Annar í Jólum",
         key: "jola2",
         holiday: true,
       },
       {
-        date: new Date(year, 11, 31),
+        date: new Date(Date.UTC(year, 11, 31)),
         description: "Gamlársdagur",
         key: "gamlars",
         holiday: true,